seoul-subway-monitor/requirements.txt
requests
supabase
python-dotenv
pandas
schedule
